third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-multiple-style-texts.html

<html>
<head>
<link rel="stylesheet" href="resources/set-style-text.css"/>
<script type="text/javascript" src="resources/inspector-protocol-test.js"></script>
<script type="text/javascript" src="resources/css-protocol-test.js"></script>
<script type="text/javascript" src="resources/dom-protocol-test.js"></script>
<script type="text/javascript">

function test()
{
    var setStyleTexts;
    var verifyProtocolError;
    var styleSheetId;
    var undoAndNext = InspectorTest.undoAndNext;

    InspectorTest.requestDocumentNodeId(onDocumentNodeId);

    function onDocumentNodeId(nodeId)
    {
        InspectorTest.eventHandler["CSS.styleSheetAdded"] = styleSheetAdded;
        InspectorTest.sendCommandOrDie("CSS.enable", {});
    }

    function styleSheetAdded(result)
    {
        styleSheetId = result.params.header.styleSheetId;
        setStyleTexts = InspectorTest.setStyleTexts.bind(InspectorTest, styleSheetId, false);
        verifyProtocolError = InspectorTest.setStyleTexts.bind(InspectorTest, styleSheetId, true);
        InspectorTest.sendCommandOrDie("CSS.getStyleSheetText", { styleSheetId: styleSheetId }, onInitialStyleSheetText);
    }

    function onInitialStyleSheetText(result)
    {
        InspectorTest.log("==== Initial style sheet text ====");
        InspectorTest.log(result.text);
        InspectorTest.runTestSuite(testSuite);
    }

    var testSuite = [
        function testMalformedArguments1(next)
        {
            verifyProtocolError([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED';\n",
                },
                {
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';\n",
                },
            ], next);
        },

        function testMalformedArguments2(next)
        {
            verifyProtocolError([
                {
                    styleSheetId: styleSheetId,
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED';\n",
                },
                {
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';\n",
                },
            ], next);
        },

        function testMalformedArguments3(next)
        {
            verifyProtocolError([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: "STRING INSTEAD OF NUMBER", startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED';\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';\n",
                },
            ], next);
        },

        function testFirstEditDoesNotApply(next)
        {
            verifyProtocolError([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED';/*\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';\n",
                },
            ], next);
        },

        function testSecondEditDoesNotApply(next)
        {
            verifyProtocolError([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED';\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';/*\n",
                },
            ], next);
        },

        function testBasicSetStyle(next)
        {
            setStyleTexts([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED';\n"
                },
            ], undoAndNext(next));
        },

        function testMultipleStyleTexts1(next)
        {
            setStyleTexts([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED1';\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED2';\n",
                },
            ], undoAndNext(next));
        },

        function testMultipleStyleTexts2(next)
        {
            setStyleTexts([
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 17, startColumn: 11, endLine: 18, endColumn: 4 },
                    text: "\n        content: 'EDITED5';\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 13, startColumn: 11, endLine: 15, endColumn: 4 },
                    text: "\n        content: 'EDITED4';\n",
                },
                {
                    styleSheetId: styleSheetId,
                    range: { startLine: 0, startColumn: 7, endLine: 2, endColumn: 0 },
                    text: "\n    content: 'EDITED3';\n",
                },
            ], undoAndNext(next));
        },
   ];
}

</script>
</head>
<body onload="runTest();">
<p>The test verifies functionality of protocol method CSS.setStyleTexts.</p>
<article id="test"></article>
</body>
</html>
